change int to double

+++ presentation/lib/fitness/widgets/goal_carousel_widget/goal_time_cal_container.dart
@@ -8,7 +8,7 @@ class GoalTimeCalContainer extends StatelessWidget {
   const GoalTimeCalContainer({super.key, required this.burnCalories, required this.durationSeconds});
 
   final int burnCalories;
-  final int durationSeconds;
+  final double durationSeconds;
 
   @override
   Widget build(BuildContext context) {
